% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_screen.R
\name{calculate_screen}
\alias{calculate_screen}
\title{Calculate screen incidence in the presence of screening.}
\usage{
calculate_screen(dset, sensitivity, attendance, screen.start.year,
  screen.stop.year)
}
\arguments{
\item{dset}{A data frame of disease incidence as produced by
\code{generate_absence} after processing by \code{calculate_clinical}.}

\item{sensitivity}{Proportion of relevant disease detected by screening.}

\item{attendance}{Proportion of individuals who attend screening tests.}

\item{screen.start.year}{Year of follow-up at which screening starts.}

\item{screen.stop.year}{Year of follow-up at which screening stops.}
}
\value{
A data frame of simulated disease incidence organized by year of
  preclinical onset, sojourn time, and year of clinical diagnosis.
}
\description{
Expand input data frame to indicate year of diagnosis and number of cases
detected by screening.
}
\details{
This function expects that the input data frame \code{dset} contains
  a variable \code{sojourn} with a single unique value. Create new variables
  to record year of each screening round (exclusive of stopping year). The
  number of false negatives are bounded (1) below by 0 and screen round - 1
  and (2) above by number of tests - 1 and sojourn time - 1 Check that
  incidence of onset matches diagnoses. Reshape dataset to indicate number of
  screen detections in each screen year.
}
\examples{
library(plyr)
library(reshape)
dset <- generate_absence(1000, 0.001, 0, 6, 10)
cset <- ddply(dset,
              .(sojourn),
              calculate_clinical,
              sensitivity=0.5,
              attendance=0.8,
              screen.start.year=0,
              screen.stop.year=10)
sset <- ddply(cset,
              .(sojourn),
              calculate_screen,
              sensitivity=0.5,
              attendance=0.8,
              screen.start.year=0,
              screen.stop.year=10)
print(head(sset))
}
\seealso{
\code{\link{calculate_clinical}}
}

