% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_overdiag.R
\name{generate_overdiag}
\alias{generate_overdiag}
\title{Simulate incidence of overdiagnosis.}
\usage{
generate_overdiag(dset, overdiag.rate)
}
\arguments{
\item{dset}{data frame of simulated population as produced by
\code{generate_absence} after processing by \code{generate_presence}.}

\item{overdiag.rate}{Proportion of screen-detected cases that are
overdiagnosed.}
}
\value{
A data frame of simulated disease incidence organized by year of
  preclinical onset, sojourn time, and year of clinical diagnosis.
}
\description{
Append to input data frame the number of overdiagnosed cases given the
specified frequency of overdiagnosis.
}
\examples{
library(plyr)
library(reshape)
dset <- generate_absence(1000, 0.001, 0, 6, 10)
dset <- generate_presence(dset, 0, 6, 0.5, 0.8, 0, 10, 10)
dset <- generate_overdiag(dset, 0.25)
print(head(dset))
}
\seealso{
\code{\link{generate_absence}}, \code{\link{generate_presence}}
}

