% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipopulation_plot.R
\name{multipopulation_plot}
\alias{multipopulation_plot}
\title{Plot incidence with gradual dissemination of screening in a population.}
\usage{
multipopulation_plot(dset, minyear = 0, maxyear = 20, line.size = 0.1,
  minyear.text = 2, maxyear.text = maxyear - 1, text.size = 3,
  text.angle = 30)
}
\arguments{
\item{dset}{A data frame as produced by \code{multipopulation_setting}.}

\item{minyear}{Minimum year to display projections.}

\item{maxyear}{Maximum year to display projections.}

\item{line.size}{Width of lines to display projections.}

\item{minyear.text}{Minimum year to display text annotation.}

\item{maxyear.text}{Maximum year to display text annotation.}

\item{text.size}{Scale of text annotation.}

\item{text.angle}{Angle of text annotation.}
}
\value{
A ggplot object.
}
\description{
Illustrate incidence projections produced as by
\code{multipopulation_setting}.
}
\examples{
library(plyr)
library(reshape)
library(ggplot2)
mpset <- multipopulation_setting()
\dontrun{
mp <- multipopulation_plot(mpset)
print(mp)
}
}
\seealso{
\code{\link{trial_plot}}
}

