% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_setting.R
\name{trial_setting}
\alias{trial_setting}
\title{Simulate incidence in a trial setting.}
\usage{
trial_setting(arm.size = 50000, onset.rate = 0.001, sojourn.min = 0,
  sojourn.max = 6, sensitivity = 0.5, attendance = 0.8,
  overdiag.rate = 0.25, screen.start.year = 1, screen.stop.year = 30,
  followup.years = 30)
}
\arguments{
\item{arm.size}{Number of individuals in each trial arm.}

\item{onset.rate}{Annual incidence rate of relevant preclinical disease.}

\item{sojourn.min}{Shortest relevant preclinical duration.}

\item{sojourn.max}{Longest relevant preclinical duration.}

\item{sensitivity}{Screen test episode sensitivity.}

\item{attendance}{Proportion of participants who attend a screen test.}

\item{overdiag.rate}{Proportion of screen detections that are overdiagnosed.}

\item{screen.start.year}{Year of follow-up at which screening starts.}

\item{screen.stop.year}{Year of follow-up at which screening stops.}

\item{followup.years}{Number of years of follow-up.}
}
\value{
A data frame of simulated disease incidence organized by year of
  preclinical onset, sojourn time, and year of diagnosis.
}
\description{
Generate a data frame representing a screening trial of \code{arm.size}
individuals in each arm and record relevant disease diagnosed with and
without screening and overdiagnoses.
}
\examples{
library(plyr)
library(reshape)
trial_incidence_3types <- function(){
    tset <- data.frame(sojourn.min=0,
                       sojourn.max=6,
                       sensitivity=0.5,
                       attendance=0.8,
                       overdiag.rate=0.25)
    tset <- rbind(transform(tset, screen.start.year=1,
                                  screen.stop.year=5,
                                  type='screen-stop'),
                  transform(tset, screen.start.year=1,
                                  screen.stop.year=30,
                                  type='screen-continue'),
                  transform(tset, screen.start.year=5,
                                  screen.stop.year=30,
                                  type='screen-continue-delay'))
    tset <- ddply(tset,
                  .(sojourn.min,
                    sojourn.max,
                    sensitivity,
                    attendance,
                    overdiag.rate,
                    screen.start.year,
                    screen.stop.year,
                    type),
                  function(x)
                  with(x, trial_setting(sojourn.min=sojourn.min,
                                        sojourn.max=sojourn.max,
                                        sensitivity=sensitivity,
                                        attendance=attendance,
                                        overdiag.rate=overdiag.rate,
                                        screen.start.year=screen.start.year,
                                        screen.stop.year=screen.stop.year))
                  )
    newscreen <- subset(tset, arm == 'screen' & type == 'screen-continue')
    newcontrol <- subset(tset, arm == 'screen' & type == 'screen-continue-delay')
    newscreen <- transform(newscreen, type='screen-continue-control-start')
    newcontrol <- transform(newcontrol, arm='control', type='screen-continue-control-start')
    tset <- subset(tset, type != 'screen-continue-delay')
    tset <- rbind(tset, newscreen, newcontrol)
    return(tset)
}
tset_3types <- trial_incidence_3types()
}
\seealso{
\code{\link{population_setting}}
}

