% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_clinical.R
\name{calculate_clinical}
\alias{calculate_clinical}
\title{Calculate clinical incidence in the presence of screening.}
\usage{
calculate_clinical(dset, sensitivity, attendance, screen.start.year,
  screen.stop.year)
}
\arguments{
\item{dset}{A data frame of disease incidence as produced by
\code{generate_absence}.}

\item{sensitivity}{Proportion of relevant disease detected by screening.}

\item{attendance}{Proportion of individuals who attend screening tests.}

\item{screen.start.year}{Year of follow-up at which screening starts.}

\item{screen.stop.year}{Year of follow-up at which screening stops.}
}
\value{
A data frame of simulated disease incidence organized by year of
  preclinical onset, sojourn time, and year of clinical diagnosis.
}
\description{
Append to input data frame number of screen tests offered and number of
interval cases not detected by screening.
}
\details{
This function expects that the input data frame \code{dset} contains
  a variable \code{sojourn} with a single unique value. The number of false
  negatives are bounded by (1) the later of onset and the start of screening
  and (2) the earlier of clinical presentation and the end of screening.
  Disease presents clinically when all screening tests are false negatives.
  Imperfect attendance is represented by removing the expected number of
  cases in individuals who attend a sensitive test scaled by the number of
  tests offered. In other words, false negative tests represent either not
  attending a test or attending a test but it does not detect latent disease.
}
\examples{
library(plyr)
library(reshape)
dset <- generate_absence(1000, 0.001, 0, 6, 10)
cset <- ddply(dset,
              .(sojourn),
              calculate_clinical,
              sensitivity=0.5,
              attendance=0.8,
              screen.start.year=0,
              screen.stop.year=10)
print(head(cset))
}
\seealso{
\code{\link{calculate_screen}}
}

