% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_screen.R
\name{calculate_screen}
\alias{calculate_screen}
\title{Calculate screen incidence in the presence of screening.}
\usage{
calculate_screen(dset, screen.start.year, screen.stop.year, attendance,
  sensitivity)
}
\arguments{
\item{dset}{A data frame of cancer incidence as produced by
\code{generate_absence} after processing by \code{calculate_clinical}.}

\item{screen.start.year}{Year of follow-up at which screening starts.}

\item{screen.stop.year}{Year of follow-up at which screening stops.}

\item{attendance}{Proportion of individuals who attend screening tests.}

\item{sensitivity}{Proportion of relevant cancers detected by screening.}
}
\value{
A data frame of simulated cancer incidence organized by year of
  preclinical onset, sojourn time, and year of clinical diagnosis.
}
\description{
Expand input data frame to indicate year of diagnosis and number of cancers
detected by screening.
}
\details{
This function expects that the input data frame \code{dset} contains
  a variable \code{sojourn} with a single unique value. Create new variables
  to record year of each screening round (exclusive of stopping year). The
  number of false negatives are bounded (1) below by 0 and screen round - 1
  and (2) above by number of tests - 1 and sojourn time - 1 Check that
  incidence of onset matches diagnoses. Reshape dataset to indicate number of
  screen detections in each screen year.
}
\examples{
library(plyr)
library(reshape)
dset <- generate_absence(1000, 10, 0.001, 0, 6)
cset <- plyr::ddply(dset,
                    "sojourn",
                    calculate_clinical,
                    screen.start.year=0,
                    screen.stop.year=10,
                    attendance=0.8,
                    sensitivity=0.5)
sset <- plyr::ddply(cset,
                    "sojourn",
                    calculate_screen,
                    screen.start.year=0,
                    screen.stop.year=10,
                    attendance=0.8,
                    sensitivity=0.5)
print(head(sset))
}
\seealso{
\code{\link{calculate_clinical}}
}

