% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_presence.R
\name{generate_presence}
\alias{generate_presence}
\title{Simulate cancer incidence in the presence of screening.}
\usage{
generate_presence(dset, followup.years, screen.start.year, screen.stop.year,
  sojourn.min, sojourn.max, attendance, sensitivity)
}
\arguments{
\item{dset}{A data frame of simulated population as produced by
\code{generate_absence}.}

\item{followup.years}{Number of years of follow-up.}

\item{screen.start.year}{Year of follow-up at which screening starts.}

\item{screen.stop.year}{Year of follow-up at which screening stops.}

\item{sojourn.min}{Minimum years of preclinical detectable period.}

\item{sojourn.max}{Maximum years of preclinical detectable period.}

\item{attendance}{Proportion of individuals who attend screening tests.}

\item{sensitivity}{Proportion of relevant cancers detected by screening.}
}
\value{
A data frame of simulated cancer incidence organized by year
  of preclinical onset, sojourn time, and year of clinical diagnosis.
}
\description{
Screen a simulated population using specified sensitivity and count screen
diagnoses in each year of screening for relevant cancers that develop in each
year with a given sojourn time.
}
\examples{
library(plyr)
library(reshape)
dset <- generate_absence(1000, 10, 0.001, 0, 6)
dset <- generate_presence(dset, 10, 0, 10, 0, 6, 0.8, 0.5)
print(head(dset))
}
\seealso{
\code{\link{generate_absence}}, \code{\link{generate_overdiag}}
}

