% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdiag.R
\docType{package}
\name{overdiag}
\alias{overdiag}
\alias{overdiag-package}
\title{overdiag: A package to investigate empirical estimates of overdiagnosis}
\description{
This package implements a deterministic model to project disease incidence in
population and trial settings. The projections can be used to examine bias in
empirical estimates of overdiagnosis and/or determine the first time point at
which empirical estimates are unbiased.
}

