% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_setting.R
\name{population_setting}
\alias{population_setting}
\title{Simulate cancer incidence in a population setting.}
\usage{
population_setting(pop.size = 1e+05, onset.rate = 0.001, sojourn.min = 0,
  sojourn.max = 6, sensitivity = 0.5, overdiag.rate = 0.25,
  screen.start.year = 4, screen.stop.year = 30, followup.years = 30)
}
\arguments{
\item{pop.size}{Number of individuals in the simulated population.}

\item{onset.rate}{Annual incidence rate of relevant preclinical cancers.}

\item{sojourn.min}{Shortest relevant preclinical duration.}

\item{sojourn.max}{Longest relevant preclinical duration.}

\item{sensitivity}{Screen test episode sensitivity.}

\item{overdiag.rate}{Proportion of screen-detected cancers that are
overdiagnosed.}

\item{screen.start.year}{Year of follow-up at which screening starts.}

\item{screen.stop.year}{Year of follow-up at which screening stops.}

\item{followup.years}{Number of years of follow-up.}
}
\value{
A data frame of simulated cancer incidence organized by year of
  preclinical onset, sojourn time, and year of diagnosis.
}
\description{
Generate a data frame representing a population of \code{pop.size}
individuals and record relevant cancers diagnosed with and without screening
and overdiagnosed cancers.
}
\examples{
library(plyr)
library(reshape)
population_incidence_3ways <- function(pop.size=1e5,
                                       screen.start.year=3){
    dset <- data.frame(sojourn.min=c(0, 0),
                       sojourn.max=c(12, 6),
                       sensitivity=0.5,
                       overdiag.rate=0.25)
    pset <- ddply(dset,
                  .(sojourn.min,
                    sojourn.max,
                    sensitivity,
                    overdiag.rate),
                  function(x)
                  with(x, population_setting(pop.size=pop.size,
                                             sojourn.min=sojourn.min,
                                             sojourn.max=sojourn.max,
                                             sensitivity=sensitivity,
                                             overdiag.rate=overdiag.rate,
                                             screen.start.year=screen.start.year))
                  )
    return(pset)
}

multipopulation_incidence_3ways <- function(pop.size=1e5,
                                            dissemination.pattern='default'){
    if(dissemination.pattern == 'default'){
        proportion <- c(0.05, 0.1, 0.15, 0.15, 0.05, 0.5)
        start.year <- c(2, 3, 4, 5, 6, 28)
        dissemination.name <- 'Cumulative uptake (years 2-6): 5\%,15\%,30\%,45\%,50\%'
    } else {
        proportion <- c(0.1, 0.3, 0.1, 0.5)
        start.year <- c(2, 3, 4, 28)
        dissemination.name <- 'Cumulative uptake (years 2-4): 10\%,40\%,50\%'
    }
    dissemination <- data.frame(pop.size=pop.size,
                                proportion=proportion,
                                start.year=start.year)
    stopifnot(with(dissemination, sum(proportion)) == 1)
    mpset <- ddply(dissemination,
                   .(proportion, start.year),
                   function(x){
                       with(x,
        population_incidence_3ways(pop.size=proportion*pop.size,
                                   screen.start.year=start.year))
                         })
    mpset <- ddply(mpset,
                   .(sojourn.min,
                     sojourn.max,
                     sensitivity,
                     overdiag.rate,
                     year),
                   summarize,
                   count_screen=sum(count_screen),
                   count_clinical=sum(count_clinical),
                   count_overdiag=sum(count_overdiag))
    mpset <- transform(mpset, dissemination=dissemination.name)
    return(mpset)
}
mpset_default <- multipopulation_incidence_3ways(dissemination.pattern='default')
mpset_variant <- multipopulation_incidence_3ways(dissemination.pattern='variant')
mpset_3ways <- rbind(mpset_default, mpset_variant)
}
\seealso{
\code{\link{trial_setting}}
}

