% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_plot.R
\name{trial_plot}
\alias{trial_plot}
\title{Plot incidence in a randomized trial.}
\usage{
trial_plot(dset, minyear = 0, maxyear = 12, maxanninc = 150,
  maxcuminc = 800)
}
\arguments{
\item{dset}{A data frame as produced by \code{multipopulation_setting}.}

\item{minyear}{Minimum year to display projections.}

\item{maxyear}{Maximum year to display projections.}

\item{maxanninc}{Maximum annual incidence to display projections.}

\item{maxcuminc}{Maximum cumulative incidence to display projections.}
}
\value{
A ggplot object.
}
\description{
Illustrate incidence projections produced as by \code{trial_setting}.
}
\examples{
library(plyr)
library(reshape)
library(ggplot2)
tset <- trial_setting()
\dontrun{
tp <- trial_plot(tset)
print(tp)
}
}
\seealso{
\code{\link{multipopulation_plot}}
}

