% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_setting.R
\name{trial_setting}
\alias{trial_setting}
\title{Simulate cancer incidence in a trial setting.}
\usage{
trial_setting(dset, arm.size = 50000, followup.years = 30,
  onset.rate = 0.001)
}
\arguments{
\item{dset}{A data frame containing \code{sojourn.min},
\code{sojourn.max}, \code{sensitivity}, and \code{overdiag.rate}.}

\item{arm.size}{Number of individuals in each trial arm.}

\item{followup.years}{Number of years of follow-up.}

\item{onset.rate}{Number of relevant preclinical cancers that develop each year.}
}
\value{
A data frame of simulated cancer incidence organized by year of
  preclinical onset, sojourn time, and year of diagnosis.
}
\description{
Generate a data frame representing a screening trial of \code{arm.size}
individuals in each arm and record relevant cancers diagnosed with and
without screening and overdiagnosed cancers.
}
\examples{
library(plyr)
library(reshape)
trial_incidence_3types <- function(){
    tset <- data.frame(sojourn.min=0,
                       sojourn.max=6,
                       sensitivity=0.5,
                       attendance=0.8,
                       overdiag.rate=0.25)
    tset <- rbind(transform(tset, screen.start.year=1,
                                  screen.stop.year=5,
                                  type='screen-stop'),
                  transform(tset, screen.start.year=1,
                                  screen.stop.year=30,
                                  type='screen-continue'),
                  transform(tset, screen.start.year=5,
                                  screen.stop.year=30,
                                  type='screen-continue-delay'))
    tset <- ddply(tset,
                  .(sojourn.min,
                    sojourn.max,
                    sensitivity,
                    attendance,
                    overdiag.rate,
                    screen.start.year,
                    screen.stop.year,
                    type),
                  trial_setting)
    tset <- transform(tset, count_clinical=ifelse(year == 0, 0, count_clinical))
    newscreen <- subset(tset, arm == 'screen' & type == 'screen-continue')
    newcontrol <- subset(tset, arm == 'screen' & type == 'screen-continue-delay')
    newscreen <- transform(newscreen, type='screen-continue-control-start')
    newcontrol <- transform(newcontrol, arm='control', type='screen-continue-control-start')
    tset <- subset(tset, type != 'screen-continue-delay')
    tset <- rbind(tset, newscreen, newcontrol)
    return(tset)
}
tset_3types <- trial_incidence_3types()
}
\seealso{
\code{\link{population_setting}}
}

